Clean up

+++ main/java/mod/sfhcore/client/renderers/Color.java
@@ -27,14 +27,14 @@ public class Color {
     public Color(int color, boolean ignoreAlpha) {
         if (ignoreAlpha) {
             this.a = 1.0f;
-            this.r = (float) (color >> 16 & 255) / 255.0F;
-            this.g = (float) (color >> 8 & 255) / 255.0F;
-            this.b = (float) (color & 255) / 255.0F;
+            this.r = (color >> 16 & 255) / 255.0F;
+            this.g = (color >> 8 & 255) / 255.0F;
+            this.b = (color & 255) / 255.0F;
         } else {
-            this.a = (float) (color >> 24 & 255) / 255.0F;
-            this.r = (float) (color >> 16 & 255) / 255.0F;
-            this.g = (float) (color >> 8 & 255) / 255.0F;
-            this.b = (float) (color & 255) / 255.0F;
+            this.a = (color >> 24 & 255) / 255.0F;
+            this.r = (color >> 16 & 255) / 255.0F;
+            this.g = (color >> 8 & 255) / 255.0F;
+            this.b = (color & 255) / 255.0F;
         }
     }
